backend/requirements.txt
fastapi
strawberry-graphql
strawberry-graphql[fastapi]
uvicorn
pillow
torch
transformers
